Conflicts:
	Content.Client/Construction/ConstructionSystem.cs
	Content.Client/Construction/UI/ConstructionMenuPresenter.cs
	Content.Client/IoC/ClientContentIoC.cs
	Content.Client/PDA/PdaMenu.xaml.cs
	Content.Client/StatusIcon/StatusIconSystem.cs
	Content.Client/Xenoarchaeology/Ui/AnalysisConsoleMenu.xaml.cs
	Content.Client/_Exodus/Mining/MiningOverlay.cs
	Content.Client/_White/Overlays/ThermalVisionOverlay.cs
	Content.IntegrationTests/Tests/PostMapInitTest.cs
	Content.Server.Database/Migrations/Sqlite/SqliteServerDbContextModelSnapshot.cs
	Content.Server/Administration/Commands/AGhostCommand.cs
	Content.Server/Administration/Commands/RoleBanCommand.cs
	Content.Server/Administration/Managers/BanManager.cs
	Content.Server/Administration/Systems/AdminVerbSystem.cs
	Content.Server/Administration/Toolshed/TagCommand.cs
	Content.Server/Chat/Managers/ChatSanitizationManager.cs
	Content.Server/Chat/Systems/ChatSystem.cs
	Content.Server/Emp/EmpSystem.cs
	Content.Server/GameTicking/GameTicker.GamePreset.cs
	Content.Server/GameTicking/GameTicker.cs
	Content.Server/GameTicking/Rules/GameRuleSystem.cs
	Content.Server/NPC/Systems/NPCCombatSystem.Ranged.cs
	Content.Server/NPC/Systems/NPCUtilitySystem.cs
	Content.Server/Projectiles/ProjectileSystem.cs
	Content.Server/Shuttles/Systems/RadarConsoleSystem.cs
	Content.Server/Shuttles/Systems/ShuttleConsoleSystem.cs
	Content.Server/Spawners/EntitySystems/ConditionalSpawnerSystem.cs
	Content.Server/Station/Commands/JobsCommand.cs
	Content.Server/Station/Systems/StationSystem.cs
	Content.Server/Weapons/Ranged/Systems/GunSystem.AutoFire.cs
	Content.Server/Xenoarchaeology/Artifact/XAE/XAETelepathicSystem.cs
	Content.Server/Xenoarchaeology/Equipment/Systems/ArtifactAnalyzerSystem.cs
	Content.Server/Xenoarchaeology/Equipment/Systems/ArtifactCrusherSystem.cs
	Content.Server/Xenoarchaeology/Equipment/Systems/NodeScannerSystem.cs
	Content.Server/Xenoarchaeology/Equipment/Systems/TraversalDistorterSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/ArtifactSystem.Actions.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/ArtifactSystem.Commands.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/ArtifactSystem.Nodes.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/ArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/ChargeBatteryArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/ChemicalPuddleArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/DamageNearbyArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/EmpArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/FoamArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/GasArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/IgniteArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/LightFlickerArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/PhasingArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/PolyOthersArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/PortalArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/RandomInstrumentArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/RandomTeleportArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/ShuffleArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/SpawnArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/TemperatureArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/ThrowArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Effects/Systems/TriggerArtifactSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactAnchorTriggerSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactDamageTriggerSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactDeathTriggerSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactElectricityTriggerSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactExamineTriggerSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactGasTriggerSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactHeatTriggerSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactInteractionTriggerSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactLandSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactMagnetTriggerSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactMicrowaveTriggerSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactMusicTriggerSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactPressureTriggerSystem.cs
	Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Systems/ArtifactTimerTriggerSystem.cs
	Content.Server/_CorvaxNext/Silicons/Borgs/AiRemoteControlSystem.cs
	Content.Server/_Crescent/ShipShields/ShipShieldsSystem.Emitter.cs
	Content.Server/_Mono/FireControl/FireControlSystem.Console.cs
	Content.Server/_Mono/MonoCoins/MonoCoinsSystem.cs
	Content.Server/_NF/Bank/BankSystem.cs
	Content.Server/_NF/Shipyard/Systems/ShipyardSystem.Consoles.cs
	Content.Shared/Humanoid/HumanoidAppearanceComponent.cs
	Content.Shared/Humanoid/SharedHumanoidAppearanceSystem.cs
	Content.Shared/Inventory/InventorySystem.Relay.cs
	Content.Shared/Physics/Controllers/SharedConveyorController.cs
	Content.Shared/Projectiles/SharedProjectileSystem.cs
	Content.Shared/Whistle/WhistleSystem.cs
	Content.Shared/Xenoarchaeology/Equipment/SuppressArtifactContainerSystem.cs
	Content.Shared/_Exodus/Mining/MiningScannerSystem.cs
	Content.Shared/_Goobstation/Clothing/Systems/SharedSealableClothingSystem.cs
	Resources/Maps/_Exodus/Shuttles/USSP/drakon.yml
	Resources/Prototypes/Entities/Objects/Specific/Janitorial/spray.yml
	Resources/Prototypes/_Mono/Catalogs/Fills/Backpacks/drone_loot.yml
	Resources/Prototypes/_Mono/Entities/Clothing/Head/modsuit.yml
	Resources/Prototypes/_Mono/Entities/Mobs/NPCs/ai.yml
	Resources/Prototypes/_Mono/Entities/SpaceArtillery/SpaceArtillery/Missile/projectiles.yml
	Resources/Prototypes/_Mono/Entities/SpaceArtillery/SpaceArtillery/Missile/tovek_launcher.yml
	Resources/Prototypes/_Mono/Entities/Structures/Machines/Computers/shipyard.yml
	Resources/Prototypes/_Mono/Shipyard/BlackMarket/hourglass.yml
	Resources/Prototypes/_Mono/Shipyard/Nfsd/ocelot.yml
	Resources/Prototypes/_Mono/Shipyard/Sr/solarsail.yml
	RobustToolbox

 It looks like you may be committing a merge.
 If this is not correct, please run
	git update-ref -d MERGE_HEAD
 and try again.

+++ Content.Server.Database/Migrations/Sqlite/SqliteServerDbContextModelSnapshot.cs
@@ -975,7 +975,7 @@ namespace Content.Server.Database.Migrations.Sqlite
                         .HasColumnType("TEXT")
                         .HasColumnName("admin_ooc_color");
 
-                    b.Property<int>("MonoCoins")
+                    b.Property<long>("MonoCoins")
                         .HasColumnType("INTEGER")
                         .HasColumnName("mono_coins");
